BT pageantと通信できるようになった

+++ make_zip.sh
@@ -1,13 +1,22 @@
 #!/bin/sh
 dest=`date +install_%y%m%d_%H%M%S`
 version=1.0alpha
-windeployqt=C:/Qt/5.9.3/msvc2017_64/bin/windeployqt.exe
-zip=pageant+_${version}_`date +%y%m%d_%H%M%S`.zip
+windeployqt=C:/Qt/5.10.0/msvc2017_64/bin/windeployqt.exe
+if [ ! -e ${windeployqt} ]; then
+	echo windeployqt not found
+	exit
+fi
+zip=pageant+_${version}_`date +%y%m%d_%H%M%S`
 if [ -e ${dest} ]; then
 	rm -rf ${dest}
+fi
+grep "^#define DEVELOP_VERSION" pageant+.h > /dev/null
+if [ "$?" -eq 0 ]; then
+	echo DEVELOP_VERSION
+	zip=${zip}_dev
 fi
 mkdir -p ${dest}/pageant+
 cp Release/pageant+.exe ${dest}/pageant+
 ${windeployqt} -no-angle --no-opengl-sw --no-translations ${dest}/pageant+/pageant+.exe
-(cd ${dest} && /bin/find . -type f | xargs -I% zip ../${zip} %)
+(cd ${dest} && /bin/find . -type f | xargs -I% zip ../${zip}.zip %)
 rm -rf ${dest}
